Appendix B - Source code/modules/LoggingScheduler/requirements.txt
azure-iot-device~=2.7.0
schedule~=1.1.0